Organise components into folders respective of their locations
