+++ package.json
@@ -18,12 +18,9 @@
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.3",
     "@types/turndown": "^5.0.1",
-<<<<<<< HEAD
     "highlight.js": "^11.5.1",
-=======
     "dexie": "^3.2.2",
     "dexie-react-hooks": "^1.1.1",
->>>>>>> feat-storage-features
     "markdown-it": "^13.0.1",
     "quill": "^1.3.7",
     "react": "^18.1.0",
