+++ src/App.tsx
@@ -3,10 +3,10 @@ import { createTheme, ThemeProvider } from '@mui/material/styles'
 import Dexie from 'dexie'
 import Quill from 'quill'
 import { ReactElement, useEffect, useState } from 'react'
-import Body from './components/Body'
-import Footer from './components/Footer'
-import Header from './components/Header'
-import Version from './components/Version'
+import Body from './components/Body/Body'
+import Footer from './components/Footer/Footer'
+import Version from './components/Footer/Version'
+import Header from './components/Header/Header'
 import toMarkdown from './converterFunctions/toMarkdown'
 
 export default function App(): ReactElement {
