+++ src/components/Footer/Version.tsx
@@ -1,4 +1,4 @@
-import packageInfo from '../../package.json'
+import packageInfo from '../../../package.json'
 
 export default () => (
     <div
