+++ src/components/Header/MenuButton.tsx
@@ -2,7 +2,7 @@ import MoreHorizIcon from '@mui/icons-material/MoreHoriz'
 import { Menu, MenuItem } from '@mui/material'
 import IconButton from '@mui/material/IconButton'
 import { useState } from 'react'
-import { toMarkdown } from '../converterFunctions'
+import { toMarkdown } from '../../converterFunctions'
 import ExportMarkdownOption from './MenuOptions/ExportMarkdownOption'
 import ThemeOption from './MenuOptions/ThemeOption'
 import UploadFileOption from './MenuOptions/UploadFileOption'
